move Server address to the client side
