+++ ServerActions/TeamActions.ts
@@ -1,5 +1,6 @@
-import { SERVER_ADDRESS, tokenFetch } from '@/components/HandballComponenets/ServerActions';
+import { tokenFetch } from '@/components/HandballComponenets/ServerActions';
 import { SearchableName, TeamStructure, TournamentStructure } from '@/ServerActions/types';
+import { SERVER_ADDRESS } from '@/app/config';
 
 interface GetTeamsArgs {
   tournament?: SearchableName;
