+++ ServerActions/TournamentActions.ts
@@ -1,5 +1,6 @@
-import { SERVER_ADDRESS, tokenFetch } from '@/components/HandballComponenets/ServerActions';
+import { tokenFetch } from '@/components/HandballComponenets/ServerActions';
 import { TournamentStructure, SearchableName } from '@/ServerActions/types';
+import { SERVER_ADDRESS } from '@/app/config';
 
 export function getTournaments(): Promise<TournamentStructure[]> {
   const url = new URL('/api/tournaments/', SERVER_ADDRESS);
