+++ app/config.ts
@@ -1,0 +1,1 @@
+export const SERVER_ADDRESS = 'https://api.squarers.club';
