+++ app/layout.tsx
@@ -3,8 +3,8 @@ import './global.css';
 
 import React from 'react';
 import { ColorSchemeScript, MantineProvider } from '@mantine/core';
-import { SERVER_ADDRESS } from '@/components/HandballComponenets/ServerActions';
 import { theme } from '@/theme';
+import { SERVER_ADDRESS } from '@/app/config';
 
 export const metadata = {
   title: 'Squarers United Sporting',
