+++ components/HandballComponenets/ServerActions.ts
@@ -1,10 +1,7 @@
 'use client';
 
 import React, { useEffect } from 'react';
-
-export const SERVER_ADDRESS = 'https://api.squarers.club';
-// export const SERVER_ADDRESS = 'http://localhost:5001/api';
-// export const SERVER_ADDRESS = 'http://49.192.26.182:5001';
+import { SERVER_ADDRESS } from '@/app/config';
 
 export async function tokenFetcher(url: string, args: any = {}) {
   const res = await tokenFetch(url, args);
